requirements.lock.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
curl_cffi==0.13.0
-e git+ssh://git@github.com/mnorodine/data_sanitizer.git@779f7558e649d68f07b1793e75d371227358e088#egg=data_sanitizer&subdirectory=../../../data_sanitizer_poo
frozendict==2.4.6
idna==3.10
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
multitasking==0.0.12
numpy==2.3.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.32.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.1.0
ruff==0.12.10
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.7
typer==0.16.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
yfinance==0.2.65
